# краулер/crawler-ecosystem/core/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: "Neural Crawler Ecosystem — Self-Evolving API"
  description: "Контракт REST API автономного самообучающегося краулера. Экосистема способна к самооптимизации, прогнозированию, A/B-тестированию стратегий и автоматическому восстановлению. Все модули взаимодействуют строго по данному контракту. Допустима кодогенерация клиентского SDK."
  version: "2.0.0"
  contact:
    name: "Core AI Team"
    email: "agi@neural-crawler.tech"
servers:
  - url: "http://localhost:3000/api"
    description: "Локальная среда разработки"
  - url: "https://api.neural-crawler.tech/api"
    description: "Продакшн (HTTPS)"
security:
  - bearerAuth: []
paths:
...
  /emails:
    get:
      tags: [Emails]
      summary: "Получить список найденных email с фильтрацией и пагинацией"
      operationId: listEmails
      parameters:
        - name: tags
          in: query
          schema: { type: array, items: { type: string } }
          explode: false
        - name: verified
          in: query
          schema: { type: boolean }
        - name: confidence_min
          in: query
          schema: { type: number, minimum: 0, maximum: 1 }
        - name: date_from
          in: query
          schema: { type: string, format: date-time }
        - name: date_to
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 1000 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: "Список email"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /emails/export:
    post:
      tags: [Emails]
      summary: "Экспортировать email в файл (JSON/CSV/XLSX)"
      operationId: exportEmails
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExportRequest" }
      responses:
        "200":
          description: "Файл экспорта"
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/json:
              schema: { type: object, properties: { downloadUrl: { type: string, format: uri } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access-токен, полученный через /auth/login"
  schemas:
...
    Email:
      type: object
      required: [id, email, sourceUrl, tags]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, pattern: '^\S+@\S+\.\S+$' }
        sourceUrl: { type: string, format: uri }
        tags: { type: array, items: { type: string } }
        confidence: { type: number }
        verified: { type: boolean }
        context: { type: string }
        lastSeen: { type: string, format: date-time }
    EmailList:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Email" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    ExportRequest:
      type: object
      required: [format]
      properties:
        format: { type: string, enum: [json, csv, xlsx] }
        filter: { $ref: "#/components/schemas/EmailFilter" }
        fields: { type: array, items: { type: string }, example: ["email", "sourceUrl", "confidence"] }
    EmailFilter:
      type: object
      properties:
        tags: { type: array, items: { type: string } }
        verified: { type: boolean }
        confidence_min: { type: number }
        date_from: { type: string, format: date-time }
        date_to: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "ERR_PROXY_FAILED" }
        message: { type: string }
        details: { type: object }
...
  responses:
...
    Unauthorized:
      description: "Не авторизован"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: "Доступ запрещён"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...